requirements.txt
requests==2.32.3
gspread==6.1.4
oauth2client==4.1.3
pywhatkit==5.4
python-dotenv==1.0.0
twilio>=8.0.0
